Map Generation is now feasible!
I created a lot of classes in order to get map generation in a state
where I can drop in an algorithm or something for creating a map on the
fly. We still need code to be added for setting a map and switching
between maps when a player steps on a trigger.