Some code is in place for Sprite Rotation
There was an attempt in mobileentity and player to put in sprite
rotation. I can't get it working properly, perhaps Ian knows what's
going on. I do have everything in place except for how to calculate the
rotation based on mouse position. If we do controller rotation, that
should be way simpler